Move `no-random-species-spawn.ts`
